Refactor defineMain to be part of /node entrypoint
