+++ code/frameworks/angular/package.json
@@ -28,6 +28,12 @@
       "require": "./dist/index.js"
     },
     "./preset": "./preset.js",
+    "./node": {
+      "types": "./dist/node/index.d.ts",
+      "node": "./dist/node/index.js",
+      "import": "./dist/node/index.js",
+      "require": "./dist/node/index.js"
+    },
     "./package.json": "./package.json"
   },
   "main": "dist/index.js",
