+++ code/frameworks/angular/src/index.ts
@@ -1,6 +1,5 @@
 export * from './client/index';
 export * from './types';
-export * from './defineMainConfig';
 
 /*
  * ATTENTION:
