+++ code/frameworks/ember/package.json
@@ -24,6 +24,12 @@
       "require": "./dist/index.js"
     },
     "./preset": "./preset.js",
+    "./node": {
+      "types": "./dist/node/index.d.ts",
+      "node": "./dist/node/index.js",
+      "import": "./dist/node/index.js",
+      "require": "./dist/node/index.js"
+    },
     "./package.json": "./package.json"
   },
   "main": "dist/index.js",
