+++ code/frameworks/ember/src/index.ts
@@ -1,7 +1,6 @@
 /// <reference types="webpack-env" />
 
 export * from './types';
-export * from './defineMainConfig';
 
 // optimization: stop HMR propagation in webpack
