+++ code/frameworks/experimental-nextjs-vite/package.json
@@ -58,6 +58,12 @@
       "import": "./dist/vite-plugin/index.mjs",
       "require": "./dist/vite-plugin/index.js"
     },
+    "./node": {
+      "types": "./dist/node/index.d.ts",
+      "node": "./dist/node/index.js",
+      "import": "./dist/node/index.mjs",
+      "require": "./dist/node/index.js"
+    },
     "./package.json": "./package.json"
   },
   "main": "dist/index.js",
@@ -134,6 +140,7 @@
       "./src/index.ts",
       "./src/vite-plugin/index.ts",
       "./src/preset.ts",
+      "./src/node/index.ts",
       "./src/preview.tsx",
       "./src/export-mocks/cache/index.ts",
       "./src/export-mocks/headers/index.ts",
