+++ code/frameworks/nextjs/package.json
@@ -87,6 +87,12 @@
       "import": "./dist/export-mocks/router/index.mjs",
       "require": "./dist/export-mocks/router/index.js"
     },
+    "./main": {
+      "types": "./dist/node/index.d.ts",
+      "node": "./dist/node/index.js",
+      "import": "./dist/node/index.mjs",
+      "require": "./dist/node/index.js"
+    },
     "./package.json": "./package.json"
   },
   "main": "dist/index.js",
@@ -209,6 +215,7 @@
       "./src/image-context.ts",
       "./src/index.ts",
       "./src/preset.ts",
+      "./src/node/index.ts",
       "./src/preview.tsx",
       "./src/export-mocks/index.ts",
       "./src/export-mocks/cache/index.ts",
