+++ code/frameworks/nextjs/src/index.ts
@@ -1,3 +1,2 @@
 export * from './types';
-export * from './defineMainConfig';
 export * from './portable-stories';
