+++ code/frameworks/react-native-web-vite/package.json
@@ -35,6 +35,12 @@
       "import": "./dist/vite-plugin.mjs",
       "require": "./dist/vite-plugin.js"
     },
+    "./node": {
+      "types": "./dist/node/index.d.ts",
+      "node": "./dist/node/index.js",
+      "import": "./dist/node/index.mjs",
+      "require": "./dist/node/index.js"
+    },
     "./package.json": "./package.json"
   },
   "main": "dist/index.js",
@@ -87,7 +93,8 @@
     "entries": [
       "./src/index.ts",
       "./src/preset.ts",
-      "./src/vite-plugin.ts"
+      "./src/vite-plugin.ts",
+      "./src/node/index.ts"
     ],
     "platform": "node"
   },
