+++ code/frameworks/react-native-web-vite/src/index.ts
@@ -1,2 +1,1 @@
 export type { FrameworkOptions, StorybookConfig } from './types';
-export * from './defineMainConfig';
