+++ code/frameworks/react-native-web-vite/src/node/index.ts
@@ -1,4 +1,4 @@
-import type { StorybookConfig } from './types';
+import type { StorybookConfig } from '../types';
 
 export function defineMain(config: StorybookConfig) {
   return config;
