+++ code/frameworks/vue3-vite/package.json
@@ -35,6 +35,12 @@
       "require": "./dist/vite-plugin.js",
       "import": "./dist/vite-plugin.mjs"
     },
+    "./node": {
+      "types": "./dist/node/index.d.ts",
+      "node": "./dist/node/index.js",
+      "import": "./dist/node/index.mjs",
+      "require": "./dist/node/index.js"
+    },
     "./package.json": "./package.json"
   },
   "main": "dist/index.js",
@@ -80,7 +86,8 @@
     "entries": [
       "./src/index.ts",
       "./src/preset.ts",
-      "./src/vite-plugin.ts"
+      "./src/vite-plugin.ts",
+      "./src/node/index.ts"
     ],
     "platform": "node"
   },
