+++ code/frameworks/vue3-vite/src/index.ts
@@ -1,2 +1,1 @@
 export * from './types';
-export * from './defineMainConfig';
