+++ code/frameworks/web-components-vite/package.json
@@ -30,6 +30,12 @@
       "types": "./dist/preset.d.ts",
       "require": "./dist/preset.js"
     },
+    "./node": {
+      "types": "./dist/node/index.d.ts",
+      "node": "./dist/node/index.js",
+      "import": "./dist/node/index.mjs",
+      "require": "./dist/node/index.js"
+    },
     "./package.json": "./package.json"
   },
   "main": "dist/index.js",
@@ -66,6 +72,7 @@
   },
   "bundler": {
     "entries": [
+      "./src/node/index.ts",
       "./src/index.ts",
       "./src/preset.ts"
     ],
