+++ code/frameworks/web-components-webpack5/package.json
@@ -33,6 +33,12 @@
       "types": "./dist/preset.d.ts",
       "require": "./dist/preset.js"
     },
+    "./node": {
+      "types": "./dist/node/index.d.ts",
+      "node": "./dist/node/index.js",
+      "import": "./dist/node/index.mjs",
+      "require": "./dist/node/index.js"
+    },
     "./package.json": "./package.json"
   },
   "main": "dist/index.js",
@@ -71,7 +77,8 @@
   "bundler": {
     "entries": [
       "./src/index.ts",
-      "./src/preset.ts"
+      "./src/preset.ts",
+      "./src/node/index.ts"
     ],
     "platform": "node"
   },
