+++ code/frameworks/web-components-webpack5/src/node/index.ts
@@ -1,4 +1,4 @@
-import { StorybookConfig } from './types';
+import type { StorybookConfig } from '../types';
 
 export function defineMain(config: StorybookConfig) {
   return config;
